+++ obj/src/persistence/IdentifiableMongoDbPersistence.d.ts
@@ -101,6 +101,10 @@ import { MongoDbPersistence } from './MongoDbPersistence';
  */
 export declare class IdentifiableMongoDbPersistence<T extends IIdentifiable<K>, K> extends MongoDbPersistence<T> implements IWriter<T, K>, IGetter<T, K>, ISetter<T> {
     /**
+     * Flag to turn on automated string ID generation
+     */
+    protected _autoGenerateId: boolean;
+    /**
      * Creates a new instance of the persistence component.
      *
      * @param collection    (optional) a collection name.